requirements.txt
pymysql==1.1.1
psycopg2-binary==2.9.11
python-dotenv==1.0.1
pytest==8.2.0
snowflake-connector-python[pandas]
dbt-snowflake==1.11.4
